Septieme foyer: 92 scripts de robustesse re-executes, 27 plateaux sur 91 changent
24 plateaux se retrecissent, 3 s'elargissent d'une cellule. Cellules validees
sur les rapports modifies: 386 -> 271 (-29.8%).

Effondrements les plus nets: cross_market_correlation_ndx_dax 27/45 -> 3/45,
dollar_neutral_composite_vol_targeted 8/9 -> 0/9,
cash_rate_correction_defensive_vol_targeting_44 3/3 -> 0/3,
delinquency_nfci_baa10y_graduated 12/15 -> 3/15.

Consequence de lecture: les mentions de 'plateau robuste' anterieures a ce
cycle ne sont pas fiables. Les rapports sont recalcules mais le texte des
entrees de backlog ne l'est pas -- en cas de divergence, results/*_robustness.md
fait foi.

2 echecs rencontres et resolus (argument de marche, dtype object sous
pandas>=3), causes deja connues sans rapport avec le bug.

Restent les 22 robustness de type portefeuille. La campagne n'est PAS close.

+++ finance/trading/scripts/nonml_multimarket_breadth_vol_targeting_overlay_robustness.py
@@ -60,7 +60,9 @@ def main():
     r = np.log(close[1:] / close[:-1])
 
     breadth = compute_multimarket_breadth_series(dates_idx)
-    breadth_raw = breadth.values
+    # astype(float) : sous pandas >= 3 la serie peut ressortir en dtype object,
+    # que np.isnan refuse. Conversion explicite, aucun changement de valeur.
+    breadth_raw = breadth.values.astype(float)
     gate_aligned = np.where(np.isnan(breadth_raw), False, breadth_raw >= BREADTH_THRESHOLD)
     first_valid = int(np.argmax(~np.isnan(breadth_raw)))
